refactor: 거리별 요금 정책 enum으로 이동.
