+++ src/test/java/nextstep/subway/line/domain/DistanceTest.java
@@ -3,12 +3,8 @@ package nextstep.subway.line.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.subway.fare.domain.Fare;
 import nextstep.subway.line.exception.IllegalDistanceException;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 class DistanceTest {
 
@@ -46,36 +42,5 @@ class DistanceTest {
             .isInstanceOf(IllegalDistanceException.class)
             .hasMessage("역과 역 사이의 거리보다 좁은 거리를 입력해주세요");
     }
-
-    @ParameterizedTest
-    @CsvSource(value = {"54:100", "59:200", "67:300", "74:300"}, delimiter = ':')
-    void 장거리의_경우_8KM까지_마다_금액이_100원씩_추가된다(int distance, int additionalFare) {
-        //when
-        Fare fare = new Distance(distance).additionalFareByDistance();
-
-        //then
-        Assertions.assertThat(fare).isEqualTo(new Fare(additionalFare));
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"14:100", "15:100", "17:200", "21:300"}, delimiter = ':')
-    void 중거리의_경우_5KM까지_마다_금액이_100원씩_추가된다(int distance, int additionalFare) {
-        //when
-        Fare fare = new Distance(distance).additionalFareByDistance();
-
-        //then
-        Assertions.assertThat(fare).isEqualTo(new Fare(additionalFare));
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"1:0", "9:0"}, delimiter = ':')
-    void 단거리의_경우_추가요금이_없다(int distance, int additionalFare) {
-        //when
-        Fare fare = new Distance(distance).additionalFareByDistance();
-
-        //then
-        Assertions.assertThat(fare).isEqualTo(new Fare(additionalFare));
-    }
-
 
 }
